% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zusmt-package.R
\docType{package}
\name{zusmt-package}
\alias{zusmt}
\alias{zusmt-package}
\title{zusmt: Bindings to the 'OpenSMT' Satisfiability Modulo Theories Solver}
\description{
Bundles the 'OpenSMT' satisfiability modulo theories (SMT) solver \url{https://github.com/usi-verification-and-security/opensmt} and exposes it to R, so that solving does not require a separately installed solver binary.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/zusmt/}
  \item Report bugs at \url{https://github.com/pedrobtz/zusmt/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com}

Authors:
\itemize{
  \item Pedro Baltazar \email{pedrobtz@gmail.com}
}

Other contributors:
\itemize{
  \item OpenSMT authors (bundled OpenSMT sources; see LICENSE.note) [contributor, copyright holder]
}

}
\keyword{internal}
